+++ packages/ui-core/CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## Unreleased
+
+## Release 1.0.13
 - **Checkbox**
     - added new component
     - added examples and description in .mdx